+++ src/types/page-tracker.ts
@@ -1,0 +1,4 @@
+import type { Pageview } from "../api/pageview";
+import type { Screenview } from "../api/screenview";
+
+export type PageTrackerParams = Pageview | Screenview;
