+++ src/types/router.ts
@@ -1,0 +1,4 @@
+export type {
+  RouteLocationNormalizedGeneric as Route,
+  Router,
+} from "vue-router";
