+++ src/types/utils.ts
@@ -1,0 +1,1 @@
+export type Tail<T> = T extends [infer _, ...infer U] ? U : never;
